requirements.txt
httpx>=0.27.0
matplotlib>=3.9.0
python-dotenv>=1.0.0
